useSettings: split out selector
